Update for Stingray version 1.1
Change to use load_project_and_scene instead of deprecated load_project.

Change to add 'false' as a third argument when removing actors from a container, which is now necessary to avoid actors from being forcefully destroyed (we will be reusing actors after removal in the example, and it will crash the engine or cause an error).